Consolidate time utils
