+++ client/verta/verta/_internal_utils/time_utils/_time_utils_py2.py
@@ -1,44 +1,23 @@
 # -*- coding: utf-8 -*-
-# pylint: disable=undefined-variable
 
-from datetime import datetime, timedelta, tzinfo
-import time
+from datetime import (  # pylint: disable=undefined-variable
+    timedelta,
+    tzinfo,
+)
 
-ZERO = timedelta(0)
 
 class UTC(tzinfo):
     """UTC"""
+    ZERO = timedelta(0)
 
     def utcoffset(self, dt):
-        return ZERO
+        return self.ZERO
 
     def tzname(self, dt):
         return "UTC"
 
     def dst(self, dt):
-        return ZERO
+        return self.ZERO
+
 
 utc = UTC()
-
-UNIX_EPOCH = datetime(1970, 1, 1, tzinfo=utc)
-
-def now():
-    return datetime.now(utc)
-
-
-def now_in_millis():
-    return int(round(time.time() * 1000))
-
-
-def epoch_millis(dt):
-    if isinstance(dt, datetime):
-        return int(round((dt - UNIX_EPOCH).total_seconds() * 1000))
-    elif type(dt) == int:
-        return dt
-    elif dt is None:
-        return dt
-    else:
-        raise ValueError("Cannot convert argument to epoch milliseconds")
-
-def datetime_from_millis(millis):
-    return UNIX_EPOCH + timedelta(milliseconds=millis)
